revisions: rebase when necessary
The rebase is done whenever the remote head diverges from the local one (i.e.
when their common ancestor is not the current head).  To avoid rewriting the
already public history, local revisions are rebased on top of the remote ones.
The delta of the latter are then applied to the local database.

This optimistically assumes that the two branches do not conflict.  If they do,
the revision graph and the database will silently diverge (since the remote
revisions come first in the graph but are applied locally last).